updated db

+++ Smartpool/Database/DatabaseModel.edmx.sql
@@ -2,8 +2,8 @@
 -- --------------------------------------------------
 -- Entity Designer DDL Script for SQL Server 2005, 2008, 2012 and Azure
 -- --------------------------------------------------
--- Date Created: 05/09/2016 12:42:05
--- Generated from EDMX file: C:\cygwin64\home\Mr. Derp-lappitoppi\git-repos\I4PRJ\Smartpool\Database\DatabaseModel.edmx
+-- Date Created: 05/09/2016 13:08:53
+-- Generated from EDMX file: C:\Users\Norgaard\Documents\Git\I4PRJ\Smartpool\Database\DatabaseModel.edmx
 -- --------------------------------------------------
 
 SET QUOTED_IDENTIFIER OFF;
@@ -89,7 +89,7 @@ GO
 -- Creating table 'pHSet'
 CREATE TABLE [dbo].[pHSet] (
     [Id] int IDENTITY(1,1) NOT NULL,
-    [Value] int  NOT NULL,
+    [Value] float  NOT NULL,
     [DataId] int  NOT NULL
 );
 GO
@@ -97,7 +97,7 @@ GO
 -- Creating table 'ChlorineSet'
 CREATE TABLE [dbo].[ChlorineSet] (
     [Id] int IDENTITY(1,1) NOT NULL,
-    [Value] int  NOT NULL,
+    [Value] float  NOT NULL,
     [DataId] int  NOT NULL
 );
 GO
@@ -105,7 +105,7 @@ GO
 -- Creating table 'TemperatureSet'
 CREATE TABLE [dbo].[TemperatureSet] (
     [Id] int IDENTITY(1,1) NOT NULL,
-    [Value] int  NOT NULL,
+    [Value] float  NOT NULL,
     [DataId] int  NOT NULL
 );
 GO
@@ -113,7 +113,7 @@ GO
 -- Creating table 'HumiditySet'
 CREATE TABLE [dbo].[HumiditySet] (
     [Id] int IDENTITY(1,1) NOT NULL,
-    [Value] int  NOT NULL,
+    [Value] float  NOT NULL,
     [DataId] int  NOT NULL
 );
 GO
